Create R Project
The confusion matrix:

The confusion matrix helps in getting the accuracy of predictivity of the model. Accuracy is one of the primary metrics of classification and prediction using a confusion matrix. Accuracy is equal to the ratio of the sum of true positives and true negatives to the Number of all assessments. The overall predictive accuracy of our attrition model is 89.45%. This means that our model could predict the employee attrition correct 89 times out of 100. This is a statistically significant result.